pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "ipython>=9.7.0",
    "opensearch-py>=2.4.0",
    "isort>=7.0.0",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
